--- image-server/src/main/java/com/wesell/imageserver/domain/ProductImage.java
@@ -1,4 +1,0 @@
-package com.wesell.imageserver.domain;
-
-public class ProductImage {
-}
