+++ image-server/src/main/java/com/wesell/imageserver/domain/dto/request/UploadFileRequestDto.java
@@ -1,0 +1,18 @@
+package com.wesell.imageserver.domain.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor(force = true, access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class UploadFileRequestDto {
+    @NotBlank
+    private Long postId;
+    @NotNull
+    private String imageUrl;
+}
