+++ image-server/src/main/java/com/wesell/imageserver/domain/entity/ProductImage.java
@@ -1,0 +1,37 @@
+package com.wesell.imageserver.domain.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@Table(name = "image")
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ProductImage {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "i_id")
+    private Long id;
+
+    @Column(name = "post_id")
+    private Long postId;
+
+    @Column(name = "i_url")
+    private String imageUrl;
+
+    @Builder
+    public ProductImage(Long postId, String imageUrl) {
+        this.postId = postId;
+        this.imageUrl = imageUrl;
+    }
+}
