+++ image-server/src/main/java/com/wesell/imageserver/domain/repository/ProductImageRepository.java
@@ -1,0 +1,7 @@
+package com.wesell.imageserver.domain.repository;
+
+import com.wesell.imageserver.domain.entity.ProductImage;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
+}
