+++ image-server/src/main/java/com/wesell/imageserver/domain/repository/ProductImageViewDao.java
@@ -1,0 +1,8 @@
+package com.wesell.imageserver.domain.repository;
+
+import com.wesell.imageserver.domain.entity.ProductImage;
+import org.springframework.data.repository.Repository;
+
+public interface ProductImageViewDao extends Repository<ProductImage, Long> {
+    ProductImage searchProductImageByPostId(Long postId);
+}
